changing total type to decimal

+++ src/db/migrations/20251110094523_sales.ts
@@ -6,7 +6,7 @@ export async function up(knex: Knex): Promise<void> {
     table.integer('customer_id').references('id').inTable('customers');
     table.integer('pharmacy_branch_id').notNullable().references('id').inTable('pharmacy_branches');
     table.integer('cashier_id').notNullable().references('id').inTable('users');
-    table.integer('total_amount').notNullable();
+    table.decimal('total_amount').notNullable();
     table.integer('payment_mode_id').notNullable().references('id').inTable('payment_modes');
     table.enum('status', ['paid', 'draft']).notNullable();
     table.timestamps(true, true);
